+++ NEWS.md
@@ -5,6 +5,8 @@
   removed.
 - VHDL-2019 mode can be enabled with `--std=2019`.  Please note there is
   very limited support for this standard at present.
+- The new `--no-save` elaboration option skips saving the elaborated
+  design and other generated files to the working library.
 
 ## Version 1.7.0 - 2022-08-07
 - *Breaking change:* In-tree builds are no longer supported: use a